# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Median Code API
  description: |
    Production-ready FastAPI code generation and API entity management with Clerk JWT authentication.

    ## Authentication

    All endpoints require Bearer JWT authentication via Clerk. Include your Clerk JWT token in the Authorization header:

    ```
    Authorization: Bearer <your-clerk-jwt-token>
    ```

    ## API Overview

    This API provides two main categories of endpoints:

    ### 1. Entity Management (CRUD Operations)

    Manage your API design entities with full CRUD operations:

    - **Namespaces** (`/namespaces`) - Create and manage namespaces for organizing entities
    - **APIs** (`/apis`) - Create and manage API definitions that group endpoints and tags
    - **Types** (`/types`) - Retrieve primitive and abstract type definitions (read-only)
    - **Validators** (`/validators`) - Retrieve validator definitions (read-only)
    - **Fields** (`/fields`) - Create, read, update, and delete field definitions
    - **Objects** (`/objects`) - Create, read, update, and delete object definitions
    - **Tags** (`/tags`) - Create, read, update, and delete endpoint tags
    - **Endpoints** (`/endpoints`) - Create, read, update, and delete API endpoint definitions

    All list endpoints support filtering by `namespaceId` query parameter.

    ### 2. Code Generation

...

    #### Generated Application Structure

    The generated zip file contains:
...
    - `requirements.txt` - Python dependencies
...

...

...

...

...
servers:
...
security:
  - BearerAuth: []
...
paths:
...
  /endpoints:
    get:
      tags:
        - Endpoints
      summary: List all endpoints
      description: Retrieve all API endpoint definitions. Endpoints can be filtered by namespace.
      operationId: listEndpoints
      parameters:
        - name: namespaceId
          in: query
          required: false
          schema:
            type: string
          description: Filter endpoints by namespace ID
      responses:
        '200':
          description: Successfully retrieved endpoints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiEndpoint'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - Endpoints
      summary: Create a new endpoint
      description: Create a new API endpoint definition.
      operationId: createEndpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - namespaceId
                - apiId
                - method
                - path
                - description
                - pathParams
                - useEnvelope
                - responseShape
              properties:
                namespaceId:
                  type: string
                  example: "namespace-user"
                apiId:
                  type: string
                  example: "api-1"
                method:
                  type: string
                  enum: [GET, POST, PUT, PATCH, DELETE]
                  example: "GET"
                path:
                  type: string
                  example: "/users/{user_id}"
                description:
                  type: string
                  example: "Get user by ID"
                tagId:
                  type: string
                  nullable: true
                  example: "tag-1"
                pathParams:
                  type: array
                  items:
                    $ref: '#/components/schemas/EndpointParameter'
                queryParamsObjectId:
                  type: string
                  nullable: true
                  example: "object-query-1"
                requestBodyObjectId:
                  type: string
                  nullable: true
                  example: "object-2"
                responseBodyObjectId:
                  type: string
                  nullable: true
                  example: "object-1"
                useEnvelope:
                  type: boolean
                  example: true
                responseShape:
                  type: string
                  enum: [object, list]
                  example: "object"
      responses:
        '201':
          description: Endpoint created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEndpoint'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /endpoints/{id}:
    get:
      tags:
        - Endpoints
      summary: Get endpoint by ID
      description: Retrieve a specific API endpoint by its ID.
      operationId: getEndpoint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Endpoint ID
      responses:
        '200':
          description: Successfully retrieved endpoint
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEndpoint'
        '404':
          description: Endpoint not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags:
        - Endpoints
      summary: Update endpoint
      description: Update an existing API endpoint definition.
      operationId: updateEndpoint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Endpoint ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                apiId:
                  type: string
                  example: "api-1"
                method:
                  type: string
                  enum: [GET, POST, PUT, PATCH, DELETE]
                  example: "POST"
                path:
                  type: string
                  example: "/users"
                description:
                  type: string
                  example: "Updated endpoint description"
                tagId:
                  type: string
                  nullable: true
                pathParams:
                  type: array
                  items:
                    $ref: '#/components/schemas/EndpointParameter'
                queryParamsObjectId:
                  type: string
                  nullable: true
                requestBodyObjectId:
                  type: string
                  nullable: true
                responseBodyObjectId:
                  type: string
                  nullable: true
                useEnvelope:
                  type: boolean
                responseShape:
                  type: string
                  enum: [object, list]
      responses:
        '200':
          description: Endpoint updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEndpoint'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Endpoint not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags:
        - Endpoints
      summary: Delete endpoint
      description: Delete an API endpoint definition.
      operationId: deleteEndpoint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Endpoint ID
      responses:
        '204':
          description: Endpoint deleted successfully
        '404':
          description: Endpoint not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk JWT token from authentication flow
  schemas:
...
    ApiEndpoint:
      type: object
      required:
        - id
        - namespaceId
        - apiId
        - method
        - path
        - description
        - pathParams
        - useEnvelope
        - responseShape
      properties:
        id:
          type: string
          description: Client-generated unique identifier
          example: "endpoint-1"
        namespaceId:
          type: string
          description: Namespace this endpoint belongs to
          example: "namespace-global"
        apiId:
          type: string
          description: Reference to the API this endpoint belongs to
          example: "api-1"
        method:
          type: string
          enum: [GET, POST, PUT, PATCH, DELETE]
          description: HTTP method
          example: "GET"
        path:
          type: string
          description: URL path with optional parameters in {braces}
          example: "/users/{user_id}"
        description:
          type: string
          description: Endpoint description for OpenAPI spec
          example: "Retrieve user by ID"
        tagId:
          type: string
          nullable: true
          description: Reference to EndpointTag.id (optional)
          example: "tag-1"
        pathParams:
          type: array
          items:
            $ref: '#/components/schemas/EndpointParameter'
          description: Path parameters extracted from URL
        queryParamsObjectId:
          type: string
          nullable: true
          description: Reference to ObjectDefinition.id for query parameters (optional)
          example: "object-query-1"
        requestBodyObjectId:
          type: string
          nullable: true
          description: Reference to ObjectDefinition.id for request body (optional - only for POST/PUT/PATCH)
          example: "object-2"
        responseBodyObjectId:
          type: string
          nullable: true
          description: Reference to ObjectDefinition.id for response body (optional)
          example: "object-1"
        useEnvelope:
          type: boolean
          description: Whether to wrap response in standard envelope (success, data, message)
          example: true
        responseShape:
          type: string
          enum: [object, list]
          description: Response data shape - single object or array of objects
          example: "object"
        expanded:
          type: boolean
          description: UI state - whether endpoint is expanded in frontend (ignored during generation)
          example: false

    # ========================================================================
    # Endpoint Parameter
    # ========================================================================
    EndpointParameter:
      type: object
      required:
        - id
        - name
        - type
        - description
        - required
      properties:
        id:
          type: string
          description: Client-generated unique identifier
          example: "param-1"
        name:
          type: string
          description: Parameter name (extracted from path {braces})
          example: "user_id"
        type:
          type: string
          description: Primitive type name
          example: "uuid"
        description:
          type: string
          description: Parameter description
          example: "Unique user identifier"
        required:
          type: boolean
          description: Whether parameter is required (path params always true)
          example: true

    # ========================================================================
    # Object Definition
    # ========================================================================
...
    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message describing what went wrong
          example: "Invalid API specification: endpoint path must start with /"
  responses:
    Unauthorized:
      description: Unauthorized - Invalid or missing Clerk JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Not authenticated"